vgn/requirements.txt
catkin_pkg
ipython
jupyter
matplotlib
numpy
open3d-python
pybullet
pylint
scipy
yapf
